+++ CHANGELOG.md
@@ -12,6 +12,7 @@ CHANGELOG
 ### Bots
 #### Collectors
 - added `intelmq.bots.parsers.opendxl.collector` (#1265).
+- added `intelmq.bots.collectors.api`: collecting data using an HTTP API (#1187).
 
 #### Parsers
 - added `intelmq.bots.parsers.mcafee.parser_atd` (#1265).